requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
PyMySQL==1.1.1
whitenoise==6.7.0
requests==2.32.3
django-allauth==0.57.0
